requirements.txt
fastapi[standard]
bs4==0.0.2
requests==2.32.3
beautifulsoup4==4.12.3
pydantic-settings==2.7.1
motor==3.6.1
validators==0.34.0
prometheus-client==0.21.1
structlog==25.1.0
pytest==8.3.4
pytest-asyncio==0.25.2
uvicorn==0.34.0
python-multipart
